chore: directory restructure
merge with directory
