+++ apps/client/src/features/app-settings/panel-utils/PanelUtils.module.scss
