+++ apps/client/src/features/app-settings/panel/about-panel/AppVersion.tsx
@@ -1,7 +1,7 @@
 import { version } from '../../../../../package.json';
 import ExternalLink from '../../../../common/components/external-link/ExternalLink';
 import useAppVersion from '../../../../common/hooks-query/useAppVersion';
-import * as Panel from '../PanelUtils';
+import * as Panel from '../../panel-utils/PanelUtils';
 
 export default function AppVersion() {
   const { data, isError } = useAppVersion();
