+++ apps/client/src/features/app-settings/panel/feature-settings-panel/FeatureSettingsPanel.tsx
@@ -1,6 +1,6 @@
 import useScrollIntoView from '../../../../common/hooks/useScrollIntoView';
 import type { PanelBaseProps } from '../../panel-list/PanelList';
-import * as Panel from '../PanelUtils';
+import * as Panel from '../../panel-utils/PanelUtils';
 
 import CustomFields from './custom-fields/CustomFields';
 import UrlPresetsForm from './UrlPresetsForm';
