+++ apps/client/src/features/app-settings/panel/project-panel/ProjectPanel.tsx
@@ -1,6 +1,6 @@
 import useScrollIntoView from '../../../../common/hooks/useScrollIntoView';
 import type { PanelBaseProps } from '../../panel-list/PanelList';
-import * as Panel from '../PanelUtils';
+import * as Panel from '../../panel-utils/PanelUtils';
 
 import ManageProjects from './ManageProjects';
 import ProjectData from './ProjectData';
